requirements.txt
numpy>=1.21
PyYAML>=6.0
